Add logging for the number of open eap-peap sessions
This change patches the freeradius server to add code to log
the number of open sessions. To make the logging less verbose it only logs changes in the number of open sessions.

Also included is a ruby file, used for testing, to create an open session. It starts a EAP_PEAP session, but then does not complete it, resulting in a zombie session.